+++ examples/svelte/counter/src/main.ts
@@ -1,0 +1,5 @@
+import App from './App.html';
+
+const app = new App({
+	target: document.querySelector('#app'),
+});