+++ examples/svelte/counter/src/typings/svelte-shims.d.ts
@@ -1,0 +1,4 @@
+declare module "*.html" {
+    // import Svelte from "svelte";
+    export default Svelte;
+}