+++ src/react/components/connect.spec.tsx
@@ -38,6 +38,86 @@ describe("redux-zero - react bindings", () => {
       expect(wrapper.html()).toBe("<h1>bye</h1>")
     })
 
+    it("should provide the actions and subscribe to changes", () => {
+      store.setState({ count: 0 })
+
+      const Comp = ({ count, increment }) => (
+        <h1 onClick={increment}>{count}</h1>
+      )
+
+      const mapToProps = ({ count }) => ({ count })
+
+      const actions = store => ({
+        increment: state => ({ count: state.count + 1 })
+      })
+
+      const ConnectedComp = connect(mapToProps, actions)(Comp)
+
+      const App = () => (
+        <Provider store={store}>
+          <ConnectedComp />
+        </Provider>
+      )
+
+      const wrapper = mount(<App />)
+
+      expect(wrapper.html()).toBe("<h1>0</h1>")
+
+      wrapper.children().simulate("click")
+      wrapper.children().simulate("click")
+
+      expect(wrapper.html()).toBe("<h1>2</h1>")
+    })
+
+    it("should peform async actions correctly", done => {
+      store.setState({ count: 0 })
+
+      const Comp = ({ count, increment }) => (
+        <h1 onClick={increment}>{count}</h1>
+      )
+
+      const mapToProps = ({ count }) => ({ count })
+
+      const actions = ({ getState, setState }) => ({
+        increment: state => {
+          Promise.resolve()
+            .then(() => {
+              setState({ pending: false, count: getState().count + 1 })
+            })
+            .then(() => {
+              setState({ count: getState().count + 1 })
+
+              const [state0, state1, state2, state3] = listener.mock.calls.map(
+                ([c]) => c
+              )
+
+              expect(state0.count).toBe(0)
+              expect(state1.pending).toBe(true)
+              expect(state1.count).toBe(0)
+              expect(state2.pending).toBe(false)
+              expect(state2.count).toBe(1)
+              expect(state3.count).toBe(2)
+
+              done()
+            })
+
+          return { pending: true }
+        }
+      })
+
+      const ConnectedComp = connect(mapToProps, actions)(Comp)
+
+      const App = () => (
+        <Provider store={store}>
+          <ConnectedComp />
+        </Provider>
+      )
+
+      const wrapper = mount(<App />)
+
+      wrapper.children().simulate("click")
+    })
+
     it("should provide the store as a prop", () => {
       const Comp = ({ store }) => <h1>{String(!!store)}</h1>
 
@@ -91,6 +171,21 @@ describe("redux-zero - react bindings", () => {
         "<div>parent bye <span>child bye</span></div>"
       )
     })
+
+    it("should connect return all state when mapToProps is not passed", () => {
+      store.setState({ message: "Hey!" })
+      const Comp = ({ message }) => <h1>{message}</h1>
+      const ConnectedComp = connect()(Comp)
+
+      const App = () => (
+        <Provider store={store}>
+          <ConnectedComp />
+        </Provider>
+      )
+
+      const wrapper = mount(<App />)
+      expect(wrapper.html()).toBe("<h1>Hey!</h1>")
+    })
   })
 
   describe("Connect component", () => {
@@ -118,6 +213,90 @@ describe("redux-zero - react bindings", () => {
       store.setState({ message: "bye" })
 
       expect(wrapper.html()).toBe("<h1>bye</h1>")
+    })
+
+    it("should provide the actions and subscribe to changes", () => {
+      store.setState({ count: 0 })
+
+      const mapToProps = ({ count }) => ({ count })
+
+      const actions = store => ({
+        increment: state => ({ count: state.count + 1 })
+      })
+
+      const ConnectedComp = () => (
+        <Connect mapToProps={mapToProps} actions={actions}>
+          {({ count, increment }) => <h1 onClick={increment}>{count}</h1>}
+        </Connect>
+      )
+
+      const App = () => (
+        <Provider store={store}>
+          <ConnectedComp />
+        </Provider>
+      )
+
+      const wrapper = mount(<App />)
+
+      expect(wrapper.html()).toBe("<h1>0</h1>")
+
+      wrapper.children().simulate("click")
+      wrapper.children().simulate("click")
+
+      expect(wrapper.html()).toBe("<h1>2</h1>")
+    })
+
+    it("should peform async actions correctly", done => {
+      store.setState({ count: 0 })
+
+      const Comp = ({ count, increment }) => (
+        <h1 onClick={increment}>{count}</h1>
+      )
+
+      const mapToProps = ({ count }) => ({ count })
+
+      const actions = ({ getState, setState }) => ({
+        increment: state => {
+          Promise.resolve()
+            .then(() => {
+              setState({ pending: false, count: getState().count + 1 })
+            })
+            .then(() => {
+              setState({ count: getState().count + 1 })
+
+              const [state0, state1, state2, state3] = listener.mock.calls.map(
+                ([c]) => c
+              )
+
+              expect(state0.count).toBe(0)
+              expect(state1.pending).toBe(true)
+              expect(state1.count).toBe(0)
+              expect(state2.pending).toBe(false)
+              expect(state2.count).toBe(1)
+              expect(state3.count).toBe(2)
+
+              done()
+            })
+
+          return { pending: true }
+        }
+      })
+
+      const ConnectedComp = () => (
+        <Connect mapToProps={mapToProps} actions={actions}>
+          {({ count, increment }) => <h1 onClick={increment}>{count}</h1>}
+        </Connect>
+      )
+
+      const App = () => (
+        <Provider store={store}>
+          <ConnectedComp />
+        </Provider>
+      )
+
+      const wrapper = mount(<App />)
+
+      wrapper.children().simulate("click")
     })
 
     it("should provide the store as a prop", () => {