+++ src/svelte/index.ts
@@ -1,0 +1,3 @@
+import { connect, getActions } from "./components/connect"
+
+export { connect, getActions }