+++ CHANGELOG.md
@@ -1,5 +1,43 @@
 # Changelog
 
+### 4.3.0
+
+- Add connect function for Svelte and usage example.
+
+### 4.2.1
+
+- Let `mapToProps` function optional
+
+If you don't pass `mapToProps` function to `connect` HOC or `Connect` component, it will inject all state as props at the connected component.
+
+```javascript
+const store = createStore({ message: 'Hey' })
+
+const App = connect()(
+  ({ message }) => (
+    <h1>{message}</h1>
+  )
+)
+```
+
+### 4.2.0
+
+- Binding actions instead of coupling them to the store.
+
+Right now, actions must import an instance of the store in order to invoke `setState()`, as discussed [here](https://github.com/concretesolutions/redux-zero/issues/16). This version solved that problem. Now it's way easier to test the actions, because they are simply pure functions:
+
+```javascript
+const createActions = store => ({
+  increment: state => ({ count: state.count + 1 }),
+})
+
+const App = connect(mapToProps, createActions)(
+  ({ count, increment }) => (
+    <button onClick={increment}>{count}</button>
+  )
+)
+```
+
 ### 4.1.1
 
 - Fixes drawback from the latest release by making `connect` HOC and `Connect` component provide the store as a prop